Cleanup setting content-type for raw POST bodies.
Recent body handling changes have (happily) started setting the content-type
header directly when encoding a body. However, in the case that the body was
simply raw bytes, and the content type was set outside of httr, that
information was being dropped. This change simply threads that information
through (and adds a test).