Cache default_ua on package load
The default_ua function is called on every request and involves
calls to the slow base function packageDescription.  Depending on the
use, this can account for up to 40% of the client-side CPU time which
is a lot when using over a loopback connection (e.g. etcd, RNeo4j).

Because this information is unlikely to change during a session (and is
infomative for the server and likely ignored anyway) computing it once
and reusing it seems safe).

By @richfitz, closes #322.